refactored references to Navbar in App.js.
