+++ server/graphiql/db/mongo/models/marketHero/index.js
@@ -4,7 +4,7 @@ import path from 'path';
 import { Promise as bbPromise } from 'bluebird';
 import marketHeroSchema from '../../schemas/marketHeroSchema';
 import db from '../../connection';
-require('dotenv').config({ path: path.resolve('.env-prod'), silent: true });
+require('dotenv').config({ path: path.resolve('.env-dev'), silent: true });
 
 /**
 * Function: "checkForLead"
