+++ src/containers/navbar/components/index.js
@@ -1,0 +1,2 @@
+export NavbarMobile from './navbar_mobile';
+export NavbarWeb from './navbar_web';
