+++ src/containers/navbar/index.js
@@ -1,1 +1,14 @@
-another PR test...
+import React from 'react';
+import {
+  NavbarWeb,
+  NavbarMobile,
+} from './components';
+
+const Navbar = () => (
+  <header className="navbar-comp-container">
+    <NavbarWeb />
+    <NavbarMobile />
+  </header>
+);
+
+export default Navbar;
