+++ tools/webpack.envs.js
@@ -1,7 +1,7 @@
 import dotenv from 'dotenv';
 import path from 'path';
 
-dotenv.load({ path: path.resolve('.env-prod') });
+dotenv.load({ path: path.resolve('.env-dev') });
 
 export default {
   development: {
